+++ crates/toolkit/tests/resolve.rs
@@ -1,4 +1,4 @@
-use farmfe_core::common::ParsedSideEffects;
+use farmfe_core::common::SideEffects;
 use farmfe_testing_helpers::fixture;
 use farmfe_toolkit::resolve::{self, PACKAGE_JSON_LOADER};
 
@@ -16,7 +16,7 @@ fn load_package_json() {
     assert_eq!(result.version, Some("1.0.0".to_string()));
     assert!(matches!(
       result.side_effects().unwrap(),
-      ParsedSideEffects::Bool(false)
+      SideEffects::Bool(false)
     ));
 
     let sub = dir.join("sub");
@@ -30,24 +30,25 @@ fn load_package_json() {
 
     assert!(matches!(
       result.side_effects().unwrap(),
-      ParsedSideEffects::Array(_)
+      SideEffects::Array(_)
     ));
 
-    if let ParsedSideEffects::Array(arr) = result.side_effects().unwrap() {
-      assert_eq!(
-        *arr,
-        vec![sub.join("main.css").to_string_lossy().to_string()]
-      );
+    if let SideEffects::Array(arr) = result.side_effects().unwrap() {
+      assert!(arr.len() == 1);
+      assert!(arr[0].is_match("main.css"));
     }
 
     // make sure cache works
     let cache = PACKAGE_JSON_LOADER.cache();
-    assert!(cache.contains_key(&main.to_string_lossy().to_string()));
-    assert!(cache.contains_key(&dir.join("src").to_string_lossy().to_string()));
-    assert!(cache.contains_key(&dir.to_string_lossy().to_string()));
+    assert!(cache.contains_key(&PACKAGE_JSON_LOADER.get_cache_key(&main, &Default::default())));
+    assert!(
+      cache.contains_key(&PACKAGE_JSON_LOADER.get_cache_key(&dir.join("src"), &Default::default()))
+    );
+    assert!(cache
+      .contains_key(&PACKAGE_JSON_LOADER.get_cache_key(&dir.to_path_buf(), &Default::default())));
 
     let cached_result = cache
-      .get(&dir.join("src").to_string_lossy().to_string())
+      .get(&PACKAGE_JSON_LOADER.get_cache_key(&dir.join("src"), &Default::default()))
       .unwrap();
     assert_eq!(
       cached_result.value().name,